requirements.txt
replicate
requests
python-dotenv